backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
colorama==0.4.6
cryptography==43.0.3
defusedxml==0.7.1
Django==4.2.20
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
dotenv==0.9.9
drf-extra-fields==3.7.0
exceptiongroup==1.2.2
filetype==1.2.0
flake8==7.1.2
gunicorn==23.0.0
idna==3.10
iniconfig==2.0.0
mccabe==0.7.0
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
pluggy==1.5.0
psycopg==3.2.6
psycopg-binary==3.2.6
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.9.0
pytest==8.3.5
python-dotenv==1.0.1
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
tomli==2.2.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0